refac: revaultd lib
expose daemon
expose revault_net and revault_tx
