+++ src/bin/cli.rs
@@ -1,4 +1,4 @@
-use common::config::{config_folder_path, Config};
+use revaultd::common::config::{config_folder_path, Config};
 
 use std::{
     env,
