+++ src/daemon/bitcoind/utils.rs
@@ -1,4 +1,4 @@
-use crate::{
+use crate::daemon::{
     bitcoind::{interface::UtxoInfo, BitcoindError},
     database::{
         interface::{
