+++ src/daemon/database/interface.rs
@@ -1,5 +1,5 @@
-use crate::{
-    assert_tx_type,
+use crate::assert_tx_type;
+use crate::daemon::{
     database::{
         schema::{
             DbSpendTransaction, DbTransaction, DbVault, DbWallet, RevaultTx, TransactionType,
