+++ src/daemon/revaultd.rs
@@ -1,4 +1,4 @@
-use common::config::{config_folder_path, BitcoindConfig, Config, ConfigError};
+use crate::common::config::{config_folder_path, BitcoindConfig, Config, ConfigError};
 
 use std::{
     collections::HashMap,
@@ -581,7 +581,7 @@ impl RevaultD {
 #[cfg(test)]
 mod tests {
     use super::RevaultD;
-    use common::config::Config;
+    use crate::common::config::Config;
 
     use std::path::PathBuf;
