+++ src/lib.rs
@@ -1,1 +1,5 @@
 pub mod common;
+pub mod daemon;
+
+pub use revault_net;
+pub use revault_tx;
